// src/app/models/character.ts
export class Character {
    id: string;
    name: string;
    bat: number;
    field: number;
    run: number;
    pitch: number;
    lean: number;
    zone: number;
    agility: number;
    coverage: number;
    reachOut: number;
    reachIn: number;
    dive: number;
    size: number;
    arm: number;
    jump: number;
    chemistry: number;
    stamina: number;
    curve: number;
    velocity: number;
    chargeVelocity: number;
    changeUp: number;

    isCaptain: boolean;
    captainAbilityOffense: string;
    captainAbilityDefense: string;

    description: string;
    notes: string;
    ability: string;
    originGroup: string;
    tags: string;
}